+++ pagination/src/main/java/com/cerner/beadledom/pagination/OffsetPaginatedListLinksWriterInterceptor.java
@@ -1,7 +1,6 @@
 package com.cerner.beadledom.pagination;
 
 import com.cerner.beadledom.pagination.models.OffsetPaginatedListDto;
-import com.cerner.beadledom.pagination.models.OffsetPaginatedListDtoImpl;
 import java.io.IOException;
 import javax.ws.rs.WebApplicationException;
 import javax.ws.rs.core.Context;
